Break up features into multiple memcache keys This avoids the 1mb memcache limit
