+++ app.yaml
@@ -1,5 +1,5 @@
 application: cr-status
-version: 2017-05-30
+version: 2017-06-05
 runtime: python27
 threadsafe: true
 api_version: 1
