+++ server.py
@@ -20,89 +20,16 @@ import logging
 import os
 import webapp2
 
-from google.appengine.api import memcache
-from google.appengine.api import urlfetch
-from google.appengine.api import users
-
 import common
 import models
 import settings
+import util
 
 import http2push.http2push as http2push
 
 
 def normalized_name(val):
   return val.lower().replace(' ', '').replace('/', '')
-
-def first_of_milestone(feature_list, milestone, start=0):
-  for i in xrange(start, len(feature_list)):
-    f = feature_list[i]
-    if (str(f['shipped_milestone']) == str(milestone) or
-        f['impl_status_chrome'] == str(milestone)):
-      return i
-    elif (f['shipped_milestone'] == None and
-          str(f['shipped_android_milestone']) == str(milestone)):
-      return i
-
-  return -1
-
-def first_of_milestone_v2(feature_list, milestone, start=0):
-  for i in xrange(start, len(feature_list)):
-    f = feature_list[i]
-    desktop_milestone = f['browsers']['chrome'].get('desktop', None)
-    android_milestone = f['browsers']['chrome'].get('android', None)
-    status = f['browsers']['chrome']['status'].get('text', None)
-
-    if (str(desktop_milestone) == str(milestone) or status == str(milestone)):
-      return i
-    elif (desktop_milestone == None and str(android_milestone) == str(milestone)):
-      return i
-
-  return -1
-
-def get_omaha_data():
-  omaha_data = memcache.get('omaha_data')
-  if omaha_data is None:
-    result = urlfetch.fetch('https://omahaproxy.appspot.com/all.json')
-    if result.status_code == 200:
-      omaha_data = json.loads(result.content)
-      memcache.set('omaha_data', omaha_data, time=86400) # cache for 24hrs.
-
-  return omaha_data
-
-def annotate_first_of_milestones(feature_list, version=None):
-  try:
-    omaha_data = get_omaha_data()
-
-    win_versions = omaha_data[0]['versions']
-
-    # Find the latest canary major version from the list of windows versions.
-    canary_versions = [x for x in win_versions if x.get('channel') and x.get('channel').startswith('canary')]
-    LATEST_VERSION = int(canary_versions[0].get('version').split('.')[0])
-
-    # TODO(ericbidelman) - memcache this calculation as part of models.py
-    milestones = range(1, LATEST_VERSION + 1)
-    milestones.reverse()
-    versions = [
-      models.IMPLEMENTATION_STATUS[models.NO_ACTIVE_DEV],
-      models.IMPLEMENTATION_STATUS[models.PROPOSED],
-      models.IMPLEMENTATION_STATUS[models.IN_DEVELOPMENT],
-      ]
-    versions.extend(milestones)
-    versions.append(models.IMPLEMENTATION_STATUS[models.NO_LONGER_PURSUING])
-
-    first_of_milestone_func = first_of_milestone
-    if version == 2:
-      first_of_milestone_func = first_of_milestone_v2
-
-    last_good_idx = 0
-    for i, ver in enumerate(versions):
-      idx = first_of_milestone_func(feature_list, ver, start=last_good_idx)
-      if idx != -1:
-        feature_list[idx]['first_of_milestone'] = True
-        last_good_idx = idx
-  except Exception as e:
-    logging.error(e)
 
 
 class MainHandler(http2push.PushHandler, common.ContentHandler, common.JSONHandler):
@@ -205,7 +132,7 @@ class MainHandler(http2push.PushHandler, common.ContentHandler, common.JSONHandl
       template_data['FEATUREOBSERVER_BUCKETS'] = json.dumps(
           properties, separators=(',',':'))
     elif path.startswith('omaha_data'):
-      omaha_data = get_omaha_data()
+      omaha_data = util.get_omaha_data()
       return common.JSONHandler.get(self, omaha_data, formatted=True)
     elif path.startswith('samples'):
       feature_list = models.Feature.get_shipping_samples() # Memcached
@@ -244,22 +171,14 @@ class MainHandler(http2push.PushHandler, common.ContentHandler, common.JSONHandl
 
 class FeaturesAPIHandler(common.JSONHandler):
 
-  def __get_feature_list(self, version=None):
-    feature_list = models.Feature.get_chronological(version=version) # Memcached
-    annotate_first_of_milestones(feature_list, version=version)
-    return feature_list
-
   def get(self, version=None):
     if version is None:
       version = 1
     else:
       version = int(version)
 
-    KEY = '%s|v%s|all' % (models.Feature.DEFAULT_MEMCACHE_KEY, version)
-    feature_list = memcache.get(KEY)
-    if feature_list is None:
-      feature_list = self.__get_feature_list(version)
-      memcache.set(KEY, feature_list)
+    feature_list = models.Feature.get_chronological(version=version) # Memcached
+
     return common.JSONHandler.get(self, feature_list, formatted=True)
